UrlSync: Export url utils
