+++ packages/scenes/src/services/UrlSyncManager.ts
@@ -1,12 +1,13 @@
 import { Location, UnregisterCallback } from 'history';
-import { isEqual } from 'lodash';
 
 import { locationService } from '@grafana/runtime';
 
 import { SceneObjectStateChangedEvent } from '../core/events';
-import { SceneObject, SceneObjectUrlValue, SceneObjectUrlValues } from '../core/types';
+import { SceneObject, SceneObjectUrlValues } from '../core/types';
 import { writeSceneLog } from '../utils/writeSceneLog';
 import { Unsubscribable } from 'rxjs';
+import { UniqueUrlKeyMapper } from './UniqueUrlKeyMapper';
+import { getUrlState, isUrlValueEqual, syncStateFromUrl } from './utils';
 
 export interface UrlSyncManagerLike {
   initSync(root: SceneObject): void;
@@ -15,7 +16,7 @@ export interface UrlSyncManagerLike {
 }
 
 export class UrlSyncManager implements UrlSyncManagerLike {
-  private urlKeyMapper = new UniqueUrlKeyMapper();
+  private _urlKeyMapper = new UniqueUrlKeyMapper();
   private _sceneRoot!: SceneObject;
   private _stateSub: Unsubscribable | null = null;
   private _locationSub?: UnregisterCallback | null = null;
@@ -71,8 +72,8 @@ export class UrlSyncManager implements UrlSyncManagerLike {
   public syncFrom(sceneObj: SceneObject) {
     const urlParams = locationService.getSearch();
     // The index is always from the root
-    this.urlKeyMapper.rebuildIndex(this._sceneRoot);
-    this._syncSceneStateFromUrl(sceneObj, urlParams);
+    this._urlKeyMapper.rebuildIndex(this._sceneRoot);
+    syncStateFromUrl(sceneObj, urlParams, this._urlKeyMapper);
   }
 
   private _onLocationUpdate = (location: Location) => {
@@ -82,9 +83,9 @@ export class UrlSyncManager implements UrlSyncManagerLike {
 
     const urlParams = new URLSearchParams(location.search);
     // Rebuild key mapper index before starting sync
-    this.urlKeyMapper.rebuildIndex(this._sceneRoot);
+    this._urlKeyMapper.rebuildIndex(this._sceneRoot);
     // Sync scene state tree from url
-    this._syncSceneStateFromUrl(this._sceneRoot, urlParams);
+    syncStateFromUrl(this._sceneRoot, urlParams, this._urlKeyMapper);
     this._lastPath = location.pathname;
   };
 
@@ -97,10 +98,10 @@ export class UrlSyncManager implements UrlSyncManagerLike {
       const searchParams = locationService.getSearch();
       const mappedUpdated: SceneObjectUrlValues = {};
 
-      this.urlKeyMapper.rebuildIndex(this._sceneRoot);
+      this._urlKeyMapper.rebuildIndex(this._sceneRoot);
 
       for (const [key, newUrlValue] of Object.entries(newUrlState)) {
-        const uniqueKey = this.urlKeyMapper.getUniqueKey(key, changedObject);
+        const uniqueKey = this._urlKeyMapper.getUniqueKey(key, changedObject);
         const currentUrlValue = searchParams.getAll(uniqueKey);
 
         if (!isUrlValueEqual(currentUrlValue, newUrlValue)) {
@@ -114,124 +115,9 @@ export class UrlSyncManager implements UrlSyncManagerLike {
     }
   };
 
-  private _syncSceneStateFromUrl(sceneObject: SceneObject, urlParams: URLSearchParams) {
-    if (sceneObject.urlSync) {
-      const urlState: SceneObjectUrlValues = {};
-      const currentState = sceneObject.urlSync.getUrlState();
-
-      for (const key of sceneObject.urlSync.getKeys()) {
-        const uniqueKey = this.urlKeyMapper.getUniqueKey(key, sceneObject);
-        const newValue = urlParams.getAll(uniqueKey);
-        const currentValue = currentState[key];
-
-        if (isUrlValueEqual(newValue, currentValue)) {
-          continue;
-        }
-
-        if (newValue.length > 0) {
-          if (Array.isArray(currentValue)) {
-            urlState[key] = newValue;
-          } else {
-            urlState[key] = newValue[0];
-          }
-        } else {
-          // mark this key as having no url state
-          urlState[key] = null;
-        }
-      }
-
-      if (Object.keys(urlState).length > 0) {
-        sceneObject.urlSync.updateFromUrl(urlState);
-      }
-    }
-
-    sceneObject.forEachChild((child) => this._syncSceneStateFromUrl(child, urlParams));
-  }
-
   public getUrlState(root: SceneObject): SceneObjectUrlValues {
-    const urlKeyMapper = new UniqueUrlKeyMapper();
-    urlKeyMapper.rebuildIndex(root);
-
-    const result: SceneObjectUrlValues = {};
-
-    const visitNode = (obj: SceneObject) => {
-      if (obj.urlSync) {
-        const newUrlState = obj.urlSync.getUrlState();
-
-        for (const [key, value] of Object.entries(newUrlState)) {
-          if (value != null) {
-            const uniqueKey = urlKeyMapper.getUniqueKey(key, obj);
-            result[uniqueKey] = value;
-          }
-        }
-      }
-
-      obj.forEachChild(visitNode);
-    };
-
-    visitNode(root);
-    return result;
+    return getUrlState(root);
   }
-}
-
-interface SceneObjectWithDepth {
-  sceneObject: SceneObject;
-  depth: number;
-}
-class UniqueUrlKeyMapper {
-  private index = new Map<string, SceneObjectWithDepth[]>();
-
-  public getUniqueKey(key: string, obj: SceneObject) {
-    const objectsWithKey = this.index.get(key);
-    if (!objectsWithKey) {
-      throw new Error("Cannot find any scene object that uses the key '" + key + "'");
-    }
-
-    const address = objectsWithKey.findIndex((o) => o.sceneObject === obj);
-    if (address > 0) {
-      return `${key}-${address + 1}`;
-    }
-
-    return key;
-  }
-
-  public rebuildIndex(root: SceneObject) {
-    this.index.clear();
-    this.buildIndex(root, 0);
-  }
-
-  private buildIndex(sceneObject: SceneObject, depth: number) {
-    if (sceneObject.urlSync) {
-      for (const key of sceneObject.urlSync.getKeys()) {
-        const hit = this.index.get(key);
-        if (hit) {
-          hit.push({ sceneObject, depth });
-          hit.sort((a, b) => a.depth - b.depth);
-        } else {
-          this.index.set(key, [{ sceneObject, depth }]);
-        }
-      }
-    }
-
-    sceneObject.forEachChild((child) => this.buildIndex(child, depth + 1));
-  }
-}
-
-export function isUrlValueEqual(currentUrlValue: string[], newUrlValue: SceneObjectUrlValue): boolean {
-  if (currentUrlValue.length === 0 && newUrlValue == null) {
-    return true;
-  }
-
-  if (!Array.isArray(newUrlValue) && currentUrlValue?.length === 1) {
-    return newUrlValue === currentUrlValue[0];
-  }
-
-  if (newUrlValue?.length === 0 && currentUrlValue === null) {
-    return true;
-  }
-
-  // We have two arrays, lets compare them
-  return isEqual(currentUrlValue, newUrlValue);
 }
 
 let urlSyncManager: UrlSyncManagerLike | undefined;
